refacto: rename handler package to web
